# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SOMELEC Mission Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /missions:
    get:
      operationId: listMissions
      tags: [missions]
      summary: List missions (scoped by user role)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: departmentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: startDateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: startDateTo
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of missions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MissionListResponse"
    post:
      operationId: createMission
      tags: [missions]
      summary: Create a new mission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMissionBody"
      responses:
        "201":
          description: Mission created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
  /missions/{id}:
    get:
      operationId: getMission
      tags: [missions]
      summary: Get mission by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Mission found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMission
      tags: [missions]
      summary: Update a mission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMissionBody"
      responses:
        "200":
          description: Mission updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
    delete:
      operationId: deleteMission
      tags: [missions]
      summary: Delete a mission (only if draft)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /missions/{id}/validate:
    post:
      operationId: validateMission
      tags: [validations]
      summary: Validate or reject a mission at current stage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateMissionBody"
      responses:
        "200":
          description: Mission validated/rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
  /missions/{id}/assign-vehicles:
    post:
      operationId: assignVehicles
      tags: [missions]
      summary: DMG assigns vehicles to a mission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignVehiclesBody"
      responses:
        "200":
          description: Vehicles assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
  /missions/{id}/generate-order:
    post:
      operationId: generateMissionOrder
      tags: [missions]
      summary: Generate mission order (CAD only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Mission order generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MissionOrder"
  /missions/{id}/order:
    get:
      operationId: getMissionOrder
      tags: [missions]
      summary: Get mission order details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Mission order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MissionOrder"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /missions/{id}/employees:
    get:
      operationId: getMissionEmployees
      tags: [missions]
      summary: Get employees assigned to a mission with fee calculations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Mission employees with fees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MissionEmployeeFee"
  /missions/{id}/validations:
    get:
      operationId: getMissionValidations
      tags: [validations]
      summary: Get validation history for a mission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Validation history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationStep"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Mission:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        needsExpression:
          type: string
        actionPlan:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        destination:
          type: string
        requiresFuel:
          type: boolean
        requiresVehicle:
          type: boolean
        vehicleCount:
          type: integer
        vehicleDetails:
          type: ["string", "null"]
        status:
          type: string
          enum:
            - draft
            - pending_director
            - pending_central_director
            - pending_technical_control
            - pending_dga
            - pending_dmg
            - pending_cad
            - pending_financial_control
            - pending_drh
            - approved
            - rejected
        currentValidationRole:
          type: ["string", "null"]
        createdByUserId:
          type: integer
        createdByName:
          type: string
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        orderNumber:
          type: ["string", "null"]
        orderGeneratedAt:
          type: ["string", "null"]
        totalFees:
          type: ["number", "null"]
        paidAmount:
          type: ["number", "null"]
        remainingAmount:
          type: ["number", "null"]
        durationDays:
          type: integer
        employeeCount:
          type: integer
        employees:
          type: array
          items:
            $ref: "#/components/schemas/MissionEmployeeSummary"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - needsExpression
        - actionPlan
        - startDate
        - endDate
        - destination
        - requiresFuel
        - requiresVehicle
        - vehicleCount
        - status
        - createdByUserId
        - createdByName
        - durationDays
        - employeeCount
        - employees
        - createdAt
        - updatedAt
    MissionEmployeeSummary:
      type: object
      properties:
        employeeId:
          type: integer
        fullName:
          type: string
        matricule:
          type: string
        position:
          type: string
        category:
          type: string
      required:
        - employeeId
        - fullName
        - matricule
        - position
        - category
    MissionEmployeeFee:
      type: object
      properties:
        employeeId:
          type: integer
        fullName:
          type: string
        matricule:
          type: string
        nni:
          type: ["string", "null"]
        position:
          type: string
        category:
          type: string
        dailyRate:
          type: number
        durationDays:
          type: integer
        totalFee:
          type: number
        paidAmount:
          type: number
        remainingAmount:
          type: number
      required:
        - employeeId
        - fullName
        - matricule
        - position
        - category
        - dailyRate
        - durationDays
        - totalFee
        - paidAmount
        - remainingAmount
    CreateMissionBody:
      type: object
      properties:
        title:
          type: string
        needsExpression:
          type: string
        actionPlan:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        destination:
          type: string
        requiresFuel:
          type: boolean
        requiresVehicle:
          type: boolean
        vehicleCount:
          type: integer
        employeeIds:
          type: array
          items:
            type: integer
      required:
        - title
        - needsExpression
        - actionPlan
        - startDate
        - endDate
        - destination
        - requiresFuel
        - requiresVehicle
        - vehicleCount
        - employeeIds
    UpdateMissionBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        needsExpression:
          type: ["string", "null"]
        actionPlan:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        destination:
          type: ["string", "null"]
        requiresFuel:
          type: ["boolean", "null"]
        requiresVehicle:
          type: ["boolean", "null"]
        vehicleCount:
          type: ["integer", "null"]
        employeeIds:
          type: array
          items:
            type: integer
    ValidateMissionBody:
      type: object
      properties:
        action:
          type: string
          enum:
            - approve
            - reject
        comment:
          type: ["string", "null"]
      required:
        - action
    AssignVehiclesBody:
      type: object
      properties:
        vehicleDetails:
          type: string
        vehicleCount:
          type: integer
      required:
        - vehicleDetails
        - vehicleCount
    MissionListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Mission"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - data
        - total
        - page
        - limit
        - totalPages
    ValidationStep:
      type: object
      properties:
        id:
          type: integer
        missionId:
          type: integer
        validatorUserId:
          type: integer
        validatorName:
          type: string
        validatorRole:
          type: string
        action:
          type: string
        comment:
          type: ["string", "null"]
        fromStatus:
          type: string
        toStatus:
          type: string
        createdAt:
          type: string
      required:
        - id
        - missionId
        - validatorUserId
        - validatorName
        - validatorRole
        - action
        - fromStatus
        - toStatus
        - createdAt
    MissionOrder:
      type: object
      properties:
        id:
          type: integer
        missionId:
          type: integer
        orderNumber:
          type: string
        missionTitle:
          type: string
        departmentName:
          type: ["string", "null"]
        destination:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        durationDays:
          type: integer
        requiresFuel:
          type: boolean
        requiresVehicle:
          type: boolean
        vehicleCount:
          type: integer
        vehicleDetails:
          type: ["string", "null"]
        employees:
          type: array
          items:
            $ref: "#/components/schemas/MissionEmployeeFee"
        totalFees:
          type: number
        paidAmount:
          type: number
        remainingAmount:
          type: number
        generatedAt:
          type: string
        generatedByName:
          type: string
      required:
        - id
        - missionId
        - orderNumber
        - missionTitle
        - destination
        - startDate
        - endDate
        - durationDays
        - requiresFuel
        - requiresVehicle
        - vehicleCount
        - employees
        - totalFees
        - paidAmount
        - remainingAmount
        - generatedAt
        - generatedByName
...