# lib/api-zod/src/generated/types/createMissionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SOMELEC Mission Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMissionBody {
  title: string;
  needsExpression: string;
  actionPlan: string;
  startDate: string;
  endDate: string;
  destination: string;
  requiresFuel: boolean;
  requiresVehicle: boolean;
  vehicleCount: number;
  employeeIds: number[];
}