# lib/api-zod/src/generated/types/validateMissionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SOMELEC Mission Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { ValidateMissionBodyAction } from "./validateMissionBodyAction";

export interface ValidateMissionBody {
  action: ValidateMissionBodyAction;
  /** @nullable */
  comment?: string | null;
}